% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi-data.R
\docType{data}
\name{data.bmi}
\alias{data.bmi}
\title{Real data for CAMRA example}
\format{
data.bmi contains the following components:
\describe{
  \item{exposure}{exposure indicator.}
  \item{mediators}{an abundance matrix of microbial taxa.}
  \item{outcome}{continuous outcome variable.}
}
}
\usage{
data(data.bmi)
}
\description{
This data is derived from `curatedMetagenomicData` R package, which includes 863 samples from China and USA.
}
\keyword{datasets}
