% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_camra.R
\docType{data}
\name{data.camra}
\alias{data.camra}
\title{Real data for CAMRA example}
\format{
data.camra contains the following components:
\describe{
  \item{treatment}{treatment indicator.}
  \item{mediators}{an abundance matrix of microbial taxa.}
  \item{outcome}{continuous outcome variable.}
}
}
\usage{
data(data.camra)
}
\description{
This data is derived from a real microbiome dataset to demonstrate the utility of the CAMRA function.
}
\keyword{datasets}
