% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTree.R
\name{prepareTree}
\alias{prepareTree}
\title{Prepare phylogenetic tree to be used in \code{\link{phyloMed}} function}
\usage{
prepareTree(tree, verbose = FALSE)
}
\arguments{
\item{tree}{A \code{phylo}-class object.}

\item{verbose}{An optional logical value. If \code{TRUE}, tree manipulation process will be printed.
Default is \code{FALSE}.}
}
\value{
A binary and rooted tree with nodes and tips conform with following \code{phylo} standards:

\itemize{
\item The internal nodes are numbered with value larger than the number of tips. 
\item The internal nodes are numbered sequentially, with values increasing away from the root.
}
}
\description{
The \code{\link{phyloMed}} function requires a rooted and binary tree as input. The \code{prepareTree} 
is an utility function that allows users to preprocess their tree if it is unrooted and/or non-binary. The tree 
output from \code{prepareTree} function can be directly fed into \code{\link{phyloMed}} function.
}
\examples{
# Load real data
data(data.cecal)
tree = prepareTree(data.cecal$tree)

}
\author{
Qilin Hong \email{qhong8@wisc.edu}
}
